Change the name of the endpoints
