+++ common/models/day-menu.js
@@ -1,6 +1,11 @@
 'use strict';
 
 module.exports = function(DayMenu) {
+  /* Validations */
+  DayMenu.validatesUniquenessOf('date', {
+    message: 'A DayMenu already exists for this date. You must remove it first'
+  });
+
   DayMenu.getMenusPerDate = function(startDate, endDate, callback) {
     DayMenu.find({
       where: {and: [
@@ -15,48 +20,11 @@ module.exports = function(DayMenu) {
     });
   };
 
-  DayMenu.publishDayMenus = function(year, weekNumber, callback) {
-    var UserMenu = DayMenu.app.models.UserMenu;
-    var DateModel = DayMenu.app.models.DateModel;
-    var today = new Date();
-    UserMenu.destroyAll({}, function(err, data) {
-      DateModel.find({
-        where: {
-          and: [
-            {year: year},
-            {week: weekNumber},
-          ],
-        },
-      }).then(function(dates) {
-        return UserMenu.find({
-        }).then(function(users) {
-          var array = dates.map(date => {
-            return users.map(user => {
-              return UserMenu.create({
-                userId: user.id,
-                status: 'O',
-                date: date.dateId,
-              }).then(function(userMenu) {
-                console.log(userMenu);
-                return userMenu;
-              });
-            });
-          });
-          console.log(array);
-          console.log(array.flat());
-          return Promise.all(array.flat());
-        }).then(function(userMenus) {
-          return userMenus;
-        });
-      }).then(function(userMenus) {
-        callback(null, userMenus);
-      });
-    });
-  };
+  /* Remote methods definition */
 
   DayMenu.remoteMethod('getMenusPerDate', {
     http: {
-      path: '/menus-per-date/:startDate/:endDate',
+      path: '/MenusPerDate/:startDate/:endDate',
       verb: 'get',
     },
     accepts: [
@@ -73,29 +41,6 @@ module.exports = function(DayMenu) {
     ],
     returns: {
       arg: 'menus',
-      type: 'array',
-    },
-  });
-
-  DayMenu.remoteMethod('publishDayMenus', {
-    http: {
-      path: '/publish-day-menus/:year/:week',
-      verb: 'get',
-    },
-    accepts: [
-      {
-        arg: 'year',
-        type: 'number',
-        required: true,
-      },
-      {
-        arg: 'week',
-        type: 'number',
-        required: true,
-      },
-    ],
-    returns: {
-      arg: 'userMenus',
       type: 'array',
     },
   });
